Add screen limits and return float variables.

+++ include/structs.h
@@ -6,7 +6,7 @@
 /*   By: msavelie <msavelie@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/09 14:20:59 by azinchen          #+#    #+#             */
-/*   Updated: 2025/05/05 16:48:40 by msavelie         ###   ########.fr       */
+/*   Updated: 2025/05/12 15:20:17 by msavelie         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -34,9 +34,9 @@ typedef enum e_obj_type {
 
 typedef struct s_vector
 {
-	double	x;
-	double	y;
-	double	z;
+	float	x;
+	float	y;
+	float	z;
 }	t_vector;
 
 typedef struct s_obj
@@ -44,16 +44,16 @@ typedef struct s_obj
 	t_obj_type	type;
 	t_vector	coordinates;
 	t_vector	normalized;
-	double		height;
-	double		diameter;
-	double		radius;
+	float		height;
+	float		diameter;
+	float		radius;
 	t_vector	color;
 	t_vector	emission_color;
 }	t_obj;
 
 typedef struct s_ambient
 {
-	double	ratio;
+	float	ratio;
 	t_color	color;
 }	t_ambient;
 
@@ -62,18 +62,18 @@ typedef struct s_light
 	t_vector	coordinates;
 	t_vector	color;
 	t_vector	emission_color;
-	double		diameter;
-	double		ratio;
+	float		diameter;
+	float		ratio;
 }	t_light;
 
 typedef struct s_camera
 {
 	t_vector	coordinates;
 	t_vector	normalized;
-	double		fov;
-	double		viewporw_size;
-	double		aspect_ratio;
-	double		angle;
+	float		fov;
+	float		viewporw_size;
+	float		aspect_ratio;
+	float		angle;
 }	t_camera;
 
 typedef struct s_ray
@@ -84,8 +84,8 @@ typedef struct s_ray
 
 typedef struct s_hit
 {
-	double		t0;
-	double		t1;
+	float		t0;
+	float		t1;
 	int			temp_part;
 	int			hit_part;
 	t_vector	phit;
@@ -98,22 +98,22 @@ typedef struct s_light_calc
 	t_vector	transmission;
 	t_vector	light_contribution;
 	t_vector	surface_color;
-	double		light_intensity;
+	float		light_intensity;
 }	t_light_calc;
 
 typedef struct s_cyl_inter
 {
 	t_vector	oc;
-	double		dir_dot_axis;
+	float		dir_dot_axis;
 	t_vector	ray_perp_dir;
-	double		ray_perp_len;
-	double		oc_dot_axis;
+	float		ray_perp_len;
+	float		oc_dot_axis;
 	t_vector	oc_proj;
-	double		discriminant;
-	double		sqrt_disc;
-	double		t0;
-	double		t1;
-	double		t_candidate;
+	float		discriminant;
+	float		sqrt_disc;
+	float		t0;
+	float		t1;
+	float		t_candidate;
 	t_vector	intersection;
 	t_vector	to_point;
 	int			*hit_part;
